Adding post detail view
- Adding PostDetailView class to the views.
- Adding post detail html.
- Adding tests for the post detail's view.
- moving fixtures from tests_posts
to conftest.py
